Create a helper to interact with hass
hass methods have gone and may go thorugh refactors. This class is a helper which will allow to decouple the interactions with hass without changing the component logic.

This helper adds the abstraction for get_url which has backward compatibility for older versions of HASS. This is the first step in solving #2