Update setup notification to include callback url
With the new get_url() method, the callback url can vary depending on which URLs are available to the user. To make the process of identifying the URL easier to add it to Oura Cloud Developer console, we are adding the callback URL into the notification. This is related to #2