Move default token file name to /data/...conf
This is only for style to move all tokens to one centralized point, similar to how `.storage` works now.

While this is a cosmetic change, it requires the user to re-authorize or move the existing token to /data/ folder and add .conf at the end. As such, we're increasing the major version as a way to indicate a breaking change.