Add async_update to avoid I/O on update
Adds an async_update wrapper which calls async_add_job to self.update. This is an easy way to make the function async and avoid blocking I/O without having to refactor major code and simply relying on Home-Assistant logic.

This should be enough to fix #6 and to have a minimal viable version for #7