Update sleep sensor to avoid overriding native attributes
Current sensor was overwriting some variables that were coming from the API and adding other values. For example, in_bed_duration was replaced from the original (in seconds) to a derived one (in hours). The purpose of this change is to ensure that users can retrieve the data exactly as the API shows it and avoid causing confusion. However, some helpful derived metrics may be added to simplify flows (e.g.  or bedtime_start_hour) or keep the sensor similar to historic needs (e.g. awake_duration_in_hours).