Refactor update method into common class
The update method on OuraDatedSensors are virtually equivalent with the exception of a couple of variables and the process to fetch and parse data. By extracting the common variables and methods, it will make adding new sensors a lot easier and help towards #10. Further refactor is possible to simplify the fetch data method.