Add heart rate sensor
This fixes #20

Heart rate belongs to a new type of sensor where instead of being timed, data is being timed. A lot of the logic from this sensor could be extracted as a base OuraDatetimedSeriesSensor derived from OuraDatedSeriesSensor. However, since there's currently only one sensor of this kind, the logic remains on the heart rate sensor instead.