Deprecate default sensor
This was not actually working since each of the sensors would load its own default and would override the default one. We could refactor the code to not load any default and get each sensor to load the defaults only if not set on default or the sensor. However, the overhead of doing that would be a lot higher and more error prone as compared to have the user to simply copy the max_backfill across all added sensors. As such, this may not be implemented back for now.