Extend end_date 1 day into the future
This is analogous to how we're extending start_date 7 days in to the past to ensure we are pulling all the data for all the dates. Now we're extending end_date one day into the future to ensure today's data is always loaded. This fixes #26 which seems to be due to these timezone differences.